+++ __tests__/__utils__/index.ts
@@ -24,9 +24,7 @@ import R from 'ramda'
 import { castToUuid, Uuid } from '~/model/decoder'
 
 export * from './assertions'
-export * from './database'
 export * from './handlers'
-export * from './error-event'
 export * from './services'
 export * from './test-client'