+++ __tests__/schema/thread/thread.ts
@@ -30,7 +30,6 @@ import {
   user,
 } from '../../../__fixtures__'
 import { given, Client, givenThreads } from '../../__utils__'
-import { encodeThreadId } from '~/schema/thread/utils'
 
 describe('uuid["threads"]', () => {
   describe('returns comment threads', () => {
@@ -75,13 +74,10 @@ describe('uuid["threads"]', () => {
             nodes: [
               {
                 comments: {
-                  nodes: [
-                    { id: encodeThreadId(comment1.id) },
-                    { id: encodeThreadId(comment2.id) },
-                  ],
+                  nodes: [{ id: comment1.id }, { id: comment2.id }],
                 },
               },
-              { comments: { nodes: [{ id: encodeThreadId(comment3.id) }] } },
+              { comments: { nodes: [{ id: comment3.id }] } },
             ],
           },
         },
@@ -94,9 +90,7 @@ describe('uuid["threads"]', () => {
       await query.shouldReturnData({
         uuid: {
           threads: {
-            nodes: [
-              { comments: { nodes: [{ id: encodeThreadId(comment3.id) }] } },
-            ],
+            nodes: [{ comments: { nodes: [{ id: comment3.id }] } }],
           },
         },
       })
@@ -126,7 +120,7 @@ describe('uuid["threads"]', () => {
                   nodes: [
                     {
                       comments: {
-                        nodes: [{ id: encodeThreadId(comment2.id) }],
+                        nodes: [{ id: comment2.id }],
                       },
                     },
                   ],
@@ -158,7 +152,7 @@ describe('uuid["threads"]', () => {
                   nodes: [
                     {
                       comments: {
-                        nodes: [{ id: encodeThreadId(comment2.id) }],
+                        nodes: [{ id: comment2.id }],
                       },
                     },
                   ],
@@ -183,9 +177,7 @@ describe('uuid["threads"]', () => {
       await query.shouldReturnData({
         uuid: {
           threads: {
-            nodes: [
-              { comments: { nodes: [{ id: encodeThreadId(comment3.id) }] } },
-            ],
+            nodes: [{ comments: { nodes: [{ id: comment3.id }] } }],
           },
         },
       })
@@ -349,6 +341,39 @@ describe('uuid["threads"]', () => {
       .shouldReturnData({
         uuid: { threads: { nodes: [{ object: { id: article.id } }] } },
       })
+  })
+
+  describe('property "object" of Comment', () => {
+    const query = new Client().prepareQuery({
+      query: gql`
+        query comments($id: Int!) {
+          uuid(id: $id) {
+            ... on Comment {
+              legacyObject {
+                id
+                alias
+              }
+            }
+          }
+        }
+      `,
+    })
+
+    test('1-level comment', async () => {
+      givenThreads({ uuid: article, threads: [[comment1]] })
+
+      await query.withVariables({ id: comment1.id }).shouldReturnData({
+        uuid: { legacyObject: { id: article.id, alias: article.alias } },
+      })
+    })
+
+    test('2-level comment', async () => {
+      givenThreads({ uuid: article, threads: [[comment1, comment2]] })
+
+      await query.withVariables({ id: comment2.id }).shouldReturnData({
+        uuid: { legacyObject: { id: article.id, alias: article.alias } },
+      })
+    })
   })
 
   test('property "createdAt" of Comment', async () => {