+++ packages/server/src/model/serlo.ts
@@ -27,7 +27,6 @@ import * as DatabaseLayer from './database-layer'
 import {
   castToUuid,
   CommentDecoder,
-  DiscriminatorType,
   EntityDecoder,
   EntityRevisionDecoder,
   NavigationDataDecoder,
@@ -59,19 +58,7 @@ export function createSerloModel({
       enableSwr: true,
       getCurrentValue: async (payload: DatabaseLayer.Payload<'UuidQuery'>) => {
         const uuid = await DatabaseLayer.makeRequest('UuidQuery', payload)
-        if (!isSupportedUuid(uuid)) return null
-        if (uuid.__typename === DiscriminatorType.User) {
-          const kratosIdentity =
-            await environment.authServices.kratos.db.getIdentityByLegacyId(
-              payload.id
-            )
-          return {
-            ...uuid,
-            language: kratosIdentity?.traits.language,
-            lastLogin: kratosIdentity?.metadata_public.lastLogin,
-          }
-        }
-        return uuid
+        return isSupportedUuid(uuid) ? uuid : null
       },
       staleAfter: { day: 1 },
       getKey: ({ id }) => {
@@ -669,24 +656,22 @@ export function createSerloModel({
     async mutate(payload: DatabaseLayer.Payload<'ThreadEditCommentMutation'>) {
       return DatabaseLayer.makeRequest('ThreadEditCommentMutation', payload)
     },
-    async updateCache(payload, value) {
-      if (value.success) {
-        await getUuid._querySpec.setCache({
-          payload: { id: payload.commentId },
-          async getValue(current) {
-            if (!current || !CommentDecoder.is(current)) return
+    async updateCache(payload) {
+      await getUuid._querySpec.setCache({
+        payload: { id: payload.commentId },
+        async getValue(current) {
+          if (!current || !CommentDecoder.is(current)) return
 
-            await getUuid._querySpec.removeCache({
-              payload: { id: current.parentId },
-            })
+          await getUuid._querySpec.removeCache({
+            payload: { id: current.parentId },
+          })
 
-            return {
-              ...current,
-              content: payload.content,
-            }
-          },
-        })
-      }
+          return {
+            ...current,
+            content: payload.content,
+          }
+        },
+      })
     },
   })